relocate sha util to auth
